FAC-242: Move AttributeLevelTarget to domain
