+++ flickit-advice-engine/src/main/java/org/flickit/assessment/advice/adapter/in/rest/advice/CreateAdviceRequestDto.java
@@ -1,6 +1,6 @@
 package org.flickit.assessment.advice.adapter.in.rest.advice;
 
-import org.flickit.assessment.advice.application.port.in.CreateAdviceUseCase.AttributeLevelTarget;
+import org.flickit.assessment.advice.application.domain.AttributeLevelTarget;
 
 import java.util.List;
