+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/application/port/in/CreateAdviceUseCaseParamTest.java
@@ -1,7 +1,7 @@
 package org.flickit.assessment.advice.application.port.in;
 
 import jakarta.validation.ConstraintViolationException;
-import org.flickit.assessment.advice.application.port.in.CreateAdviceUseCase.AttributeLevelTarget;
+import org.flickit.assessment.advice.application.domain.AttributeLevelTarget;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
